Separate Component from index.jsx.
Create each Component file to separate component from index.jsx.
