+++ src/Page.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Counter from './Counter';
+import Buttons from './Buttons';
+
+function Page({ count, onClick }) {
+  return (
+    <div>
+      <p>Counter</p>
+      <Counter count={count} onClick={() => onClick(1)} />
+      <Buttons onClick={onClick} />
+    </div>
+  );
+}
+
+export default Page;
